Change package structure.
