+++ src/main/java/oripa/domain/fold/KawasakiTheorem.java
@@ -23,7 +23,7 @@ import javax.vecmath.Vector2d;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import oripa.util.collection.AbstractRule;
+import oripa.util.rule.AbstractRule;
 import oripa.value.OriLine;
 
 /**
