+++ src/main/java/oripa/util/rule/AbstractRule.java
@@ -16,7 +16,7 @@
     You should have received a copy of the GNU General Public License
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package oripa.util.collection;
+package oripa.util.rule;
 
 /**
  * @author Koji
