+++ src/main/java/oripa/util/rule/SingleRuleConjunction.java
@@ -16,34 +16,36 @@
     You should have received a copy of the GNU General Public License
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package oripa.util.collection;
+package oripa.util.rule;
 
 import java.util.Collection;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 /**
  * @author Koji
  *
  */
-public class ConjunctionLoop<Variable> extends AbstractRule<Collection<Variable>> {
+public class SingleRuleConjunction<Variable> extends AbstractRule<Collection<Variable>> {
 
-	private final Rule<Variable> term;
+	private final Rule<Variable> rule;
 
 	/**
 	 *
-	 * @param term
+	 * @param rule
 	 */
-	public ConjunctionLoop(final Rule<Variable> term) {
-		this.term = term;
+	public SingleRuleConjunction(final Rule<Variable> rule) {
+		this.rule = rule;
 	}
 
 	@Override
 	public boolean holds(final Collection<Variable> inputs) {
-		return inputs.stream().allMatch(input -> term.holds(input));
+		return inputs.stream().allMatch(input -> rule.holds(input));
 	}
 
-	public Collection<Variable> findViolations(final Collection<Variable> inputs) {
-		CollectionFilter<Variable> filter = new CollectionFilter<>(term);
-
-		return filter.findViolations(inputs);
+	public Set<Variable> findViolations(final Collection<Variable> inputs) {
+		return inputs.stream()
+				.filter(input -> rule.violates(input))
+				.collect(Collectors.toSet());
 	}
 }
